hf-space/requirements.txt
# Backend dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Gemini (FREE tier)
google-generativeai==0.8.3

# OpenAI (optional, for RAG)
openai==1.12.0

# Vector database (optional, for RAG)
qdrant-client==1.7.0

# Utilities
pydantic-settings==2.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
